Set filter_users for NSS service in sssd.conf
A bug in path completion for Bash results in NSS requests for the "*"
user. When TACACS+ is configured this causes spurious authorization
requests for "*" when path completion is requested by a user in their
shell.

Since an asterisk is not a valid username we configure the SSSD NSS
service to ignore NSS requests for this user, suppressing the TACACS+
authorization requests.

This is a workaround for Debian bug #825317.

VRVDR-49750