vyatta-sssd: code injection due to Local.pm module
Avoid risk of code injection due to the user and group arguments passed
when calling sssd commands in the Local.pm module. Handle these as
literal strings by enclosing them in single quotes, which stops any
shell expansion for the commands. Note that the variables are still
interpolated by the script. While in most cases these arguments have
been sanitised by the caller, this might not always be the case.

VRVDR-49707