Disable NSS local_negative_timeout
Since SSSD 1.14.1 local users are held in the negative cache, by
default, for 4 hours. When there is a negative cache hit we don't
query the data provider, which means we can't determine whether it
is offline, which results in the last resort local user fallback
being unreliable.

The local_negative_timeout config for the NSS service controls this
new behaviour. Setting it to 0 disables it and causes local users to
be negatively cached for the existing entry_negative_timeout period,
which defaults to 15 seconds.

VRVDR-46487