SSSD: propagate errors from tacplus dom providers
Turn on the propagate_provider_error option for the
vyatta_system_tacplus domain.

This option was added in the DANOS version of SSSD 1.16 and is
required to allow local user login to be permitted only when TACACS+
servers are not responding (in the case where TACACS+ authentication
is being enforced, ie. only tacplus is present in the auth-chain
configuration).