Add 10-sssd cli-sandbox post create hook script
The post create hook script is run after a sandbox environment has
been created on disk, but not yet initialised or started.

It creates the necessary systemd.nspawn bind mount configuration
required for SSSD to function in the sandboxes of TACACS+ users.

Previously this was hardcoded or handled directly by the cli-sandbox
script (which now executes this hook).

The hook is installed by a new vyatta-sssd-cli-sandbox package.

VRDVR-52842