Only mount privileged PAM pipe in sandboxes
Now that cli-sandbox generates a dedicated /etc/passwd and /etc/group
for each sandbox, there is no longer any need to communicate with
SSSD for identity purposes from within a sandbox.

The only interaction with SSSD which is still required is to close
the PAM session which generates a TACACS+ session accounting stop
request. Since the (PAM interacting) login processes run as root,
the only pipe necessary for this the privileged one.

Therefore the cli-sandbox post create hook is updated to bind mount
only the private PAM pipe. Since NSS lookups are no longer necessary,
/etc/nsswitch.conf is no longer mounted in sandboxes.

Additionally add a cli-sandbox init hook which ensures restrictive
permissions on the parent directory ("private") of the pipe.
systemd-nspawn will initially create the directory with mode 755,
which the hook changes to 700 as this matches the mode on the host
filesystem.

VRVDR-52842